+++ didik/Model/ViewType.swift
@@ -1,0 +1,14 @@
+//
+//  ViewType.swift
+//  didik
+//
+//  Created by Fandrian Rhamadiansyah on 13/11/20.
+//
+
+import Foundation
+
+enum ViewType: String {
+    case jelajah = "Jelajah Materi"
+    case myMateri = "Materi Saya"
+    case lihatSemua
+}
